requirements.txt
flask
mcp
groq
requests
python-dotenv
